% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Base_grupo5}
\alias{Base_grupo5}
\title{Dados Socioeconômicos e Climáticos Municipais}
\format{

}
\usage{
Base_grupo5
}
\description{
Dataset longitudinal com indicadores municipais brasileiros (1990-2010)
}
\details{
Contém dados climáticos (temperaturas) e socioeconômicos para três décadas,
permitindo análises temporais e espaciais. Valores NA indicam dados não disponíveis
para determinado ano/município.
}
\note{
Dados compilados de fontes públicas (Atlas Brasil, INMET, etc.)
}
\keyword{datasets}
