% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidade.R
\name{carregar_dados_completos}
\alias{carregar_dados_completos}
\title{Carregar dados municipais atualizados}
\usage{
carregar_dados_completos(ano = 2010)
}
\arguments{
\item{ano}{Ano de referência para shapes}
}
\description{
Carregar dados municipais atualizados
}
