% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidade.R
\name{carregar_municipios}
\alias{carregar_municipios}
\title{Carrega dados de municípios do IBGE}
\usage{
carregar_municipios(ano = 2010)
}
\arguments{
\item{ano}{Ano dos dados (padrão: 2010)}
}
\value{
Objeto sf com shapes municipais
}
\description{
Carrega dados de municípios do IBGE
}
