% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_mapa.R
\name{criar_mapa}
\alias{criar_mapa}
\title{Cria mapa temático}
\usage{
criar_mapa(dados, variavel, titulo, paleta = "D")
}
\arguments{
\item{dados}{Dados espaciais (sf)}

\item{variavel}{String com o nome da variável a mapear}

\item{titulo}{Título do mapa}

\item{paleta}{Opção de paleta do viridis}
}
\description{
Cria mapa temático
}
