% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniao_map}
\alias{uniao_map}
\title{Dados municipais unificados}
\format{
Um objeto sf (simple features) com:
\describe{
  \item{colunas}{Todas as variáveis de Base_grupo5}
  \item{geom}{Geometria dos municípios (polígonos)}
}
}
\usage{
uniao_map
}
\description{
Combinação de dados socioeconômicos e geometrias municipais
}
\keyword{datasets}
